+++ library/Episciences/Paper/File.php
@@ -1,5 +1,8 @@
 <?php
 
+/**
+ * @property string $_downloadLike
+ */
 class Episciences_Paper_File
 {
 
@@ -161,12 +164,12 @@ class Episciences_Paper_File
     }
 
     /**
-     * @param string $fileType
+     * @param string|null $fileType
      * @return Episciences_Paper_File
      */
-    public function setFileType(string $fileType): Episciences_Paper_File
+    public function setFileType(string $fileType = null): Episciences_Paper_File
     {
-        $this->_fileType = $fileType;
+        $this->_fileType = $fileType ?? 'pdf';
         return $this;
     }
 
@@ -197,12 +200,12 @@ class Episciences_Paper_File
     }
 
     /**
-     * @param string $checksumType
+     * @param string|null $checksumType
      * @return Episciences_Paper_File
      */
-    public function setChecksumType(string $checksumType): Episciences_Paper_File
+    public function setChecksumType(string $checksumType = null): Episciences_Paper_File
     {
-        $this->_checksumType = $checksumType;
+        $this->_checksumType = $checksumType ?? 'MD5';
         return $this;
     }
 
@@ -215,12 +218,12 @@ class Episciences_Paper_File
     }
 
     /**
-     * @param string $link
+     * @param string|null $link
      * @return Episciences_Paper_File
      */
-    public function setSelfLink(string $link): Episciences_Paper_File
+    public function setSelfLink(string $link = null): Episciences_Paper_File
     {
-        $this->_selfLink = $link;
+        $this->_selfLink = $link ?? '#';
         return $this;
     }
 
@@ -241,6 +244,18 @@ class Episciences_Paper_File
     {
         $this->_timeModified = new DateTime($timeModified);
         return $this;
+    }
+
+    public function getName(): string
+    {
+        $name = $this->getFileName();
+
+        if ($this->getFileType()) {
+            $name .= '.';
+            $name .=  $this->getFileType();
+        }
+
+        return $name;
     }
 
 }