+++ library/Episciences/Tools.php
@@ -1661,4 +1661,12 @@ class Episciences_Tools
     {
         return explode('@', $string, 3);
     }
+
+    public static function isDoi(string $doi = '') : bool
+    {
+        return !($doi === '' || !preg_match("/^10.\d{4,9}\/[-._;()\/:A-Z0-9]+$/i", $doi));
+    }
+
+
+
 }