+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Security
 -->
+### fixed
+- Zenodo submissions is out of service following the update of 13/10/2023.
 
 ## 1.0.40.18 - 2023-10-19
 ## Changed